Fixed longstanding issue with trading cards' level, added status message for dev mode.
Trading cards, when added to a user's collection, now properly combine with other cards of the same name and level. For example having a collection of the same cards with +1, +0, and drawing another copy of the same card (a +0 level) will combine the two +0's into a +1. Now we have two +1's and the algorithm will take those two and combine them into a single +2.